Move exceptions builder to lists plugin (#94002)
