+++ x-pack/plugins/security_solution/common/detection_engine/types.ts
@@ -56,21 +56,3 @@ export interface EqlSearchResponse<T> {
     events?: Array<BaseHit<T>>;
   };
 }
-
-export interface BooleanFilter {
-  bool: {
-    must?: unknown | unknown[];
-    must_not?: unknown | unknown[];
-    should?: unknown[];
-    filter?: unknown | unknown[];
-    minimum_should_match?: number;
-  };
-}
-
-export interface NestedFilter {
-  nested: {
-    path: string;
-    query: unknown | unknown[];
-    score_mode: string;
-  };
-}
